apps/worker/requirements.txt
flask==3.0.0
pandas>=2.2.0
yfinance>=0.2.36
supabase>=2.9.0
python-dotenv==1.0.1
apscheduler==3.10.4
httpx>=0.27.0
httpcore>=1.0.0
statsmodels>=0.14.0
scipy>=1.10.0
pytest==8.0.0
flake8>=7.0.0